requirements.txt
aioredis==1.3.1
amqp==5.0.6
async-timeout==3.0.1
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dnspython==1.16.0
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
hiredis==2.0.0
httpcore==0.12.3
httpx==0.17.1
idna==2.10
kombu==5.0.2
motor==2.3.1
passlib==1.7.4
Pillow==8.2.0
prompt-toolkit==3.0.18
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
pymongo==3.11.3
python-jose==3.2.0
python-Levenshtein==0.12.2
python-slugify==5.0.2
pytz==2021.1
requests==2.25.1
rfc3986==1.4.0
rsa==4.7.2
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
text-unidecode==1.3
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
vine==5.0.0
wcwidth==0.2.5